+++ kubestash/__init__.py
@@ -8,6 +8,9 @@ import os
 import kubernetes
 import credstash
 import boto3
+import copy
+import traceback
+from collections import namedtuple
 
 
 # TODO: args.profile, args.arn
@@ -46,33 +49,11 @@ def base_parser():
                         help='For SECRET keys, lowercase and convert "_" to "-" (DNS_SUBDOMAIN). '
                              'Useful for compatibility with older Kubernetes versions. '
                              '(DEPRECATED).')
-    return parser
-
-
-def add_parser_inject(parent):
-    """ Parses arguments for the inject command. """
-    parser = parent.add_parser('inject',
-                               parents=[base_parser()],
-                               help='inject env variables into a Kubernetes deployment manifest, '
-                                    'taken from a Kubernetes secret (DEPRECATED; see README.md, use envFrom instead)')
-    parser.add_argument('secret',
-                        action='store',
-                        type=str,
-                        help='Kubernetes secret you want to take values from')
-    parser.add_argument('deployment',
-                        action='store',
-                        type=str,
-                        help='Kubernetes deployment to inject env values into')
-    parser.add_argument('-c', '--container',
-                        action='append',
-                        default=[],
-                        type=str,
-                        help='specify one or more containers to insert env values into (default is all containers)')
-    parser.add_argument('-u', '--update-only',
-                        dest='update_only',
+    parser.add_argument('-U', '--uppercase',
+                        dest='uppercase',
                         action='store_true',
-                        help='only update envs that are already present (do not append), useful for updating envs to '
-                             'point to a different secret')
+                        help='For environment variable keys, uppercase and convert dashes to undescores.'
+                        'Useful if your keys in credstash are in lowercase')
     return parser
 
 
@@ -94,7 +75,41 @@ def add_parser_push(parent):
                         action='store',
                         type=str,
                         default=None,
-                        help='kubernetes context')
+                        help='kubernetes context (ignored if proxy is set)')
+    parser.add_argument('-r', '--region',
+                        dest='region',
+                        action='store',
+                        type=str,
+                        default='us-east-1',
+                        help='aws region')
+    return parser
+
+
+def add_parser_pushall(parent):
+    """ Parses arguments for the pushall command. """
+    parser = parent.add_parser('pushall',
+                               parents=[base_parser()],
+                               help='push values from a Credstash table to a Kubernetes cluster')
+    parser.add_argument('table',
+                        action='store',
+                        type=str,
+                        help='Credstash table you want to pull values from')
+
+    parser.add_argument('--secretname',
+                        action='store',
+                        type=str,
+                        help='ENV_NAME you want to sync (requires --secret and --namespace)')
+
+    parser.add_argument('--secret',
+                        action='store',
+                        type=str,
+                        help='Kubernetes secret you want to push values in')
+    parser.add_argument('-c', '--context',
+                        dest='context',
+                        action='store',
+                        type=str,
+                        default=None,
+                        help='kubernetes context (ignored if proxy is set)')
     parser.add_argument('-r', '--region',
                         dest='region',
                         action='store',
@@ -124,7 +139,38 @@ def add_parser_daemon(parent):
                         action='store',
                         type=str,
                         default=None,
-                        help='kubernetes context')
+                        help='kubernetes context (ignored if proxy is set)')
+    parser.add_argument('-r', '--region',
+                        dest='region',
+                        action='store',
+                        type=str,
+                        default=None,
+                        help='aws region')
+    parser.add_argument('-i', '--interval',
+                        dest='interval',
+                        action='store',
+                        type=int,
+                        default=10,
+                        help='how long to sleep between shard iterations (seconds)')
+    return parser
+
+
+def add_parser_daemonall(parent):
+    """ Parses arguments for the daemon command. """
+    parser = parent.add_parser('daemonall',
+                               parents=[base_parser()],
+                               help='daemon mode; automatically syncs your credstash table with your entire cluster. Requires DynamoDB Streams to be enabled for your table. '
+                                    'Implies -f --force.')
+    parser.add_argument('table',
+                        action='store',
+                        type=str,
+                        help='Credstash table you want to pull values from')
+    parser.add_argument('-c', '--context',
+                        dest='context',
+                        action='store',
+                        type=str,
+                        default=None,
+                        help='kubernetes context (ignored if proxy is set)')
     parser.add_argument('-r', '--region',
                         dest='region',
                         action='store',
@@ -150,9 +196,10 @@ def parse_args():
     parsers = parser.add_subparsers(dest='cmd')
     parsers.required = True
 
-    add_parser_inject(parsers)
     add_parser_push(parsers)
+    add_parser_pushall(parsers)
     add_parser_daemon(parsers)
+    add_parser_daemonall(parsers)
 
     args = parser.parse_args()
 
@@ -173,6 +220,17 @@ def credstash_getall(args):
     return secrets
 
 
+def credstash_getone(name, args):
+    """ Returns one single secret from from credstash table `args.table`. """
+    if args.verbose:
+        print('fetching your secret from "{table}" '.format(table=args.table))
+    session_params = credstash.get_session_params(None, None)
+    return credstash.getSecret(name,
+                               region=args.region,
+                               table=args.table,
+                               **session_params)
+
+
 def dns_subdomain(string):
     """
     Converts an ENV_VARIABLE style string to a secret-style string.
@@ -189,9 +247,14 @@ def dns_subdomain(string):
     return string.replace('_', '-').lower()
 
 
-def maybe_dns_subdomain(args, string):
-    """Only convert to dns_subdomain if the --lowercase flag is set. """
-    return dns_subdomain(string) if args.lowercase else string
+def generate_key(args, string):
+    """Only convert to dns_subdomain if the --lowercase flag is set.
+       Only convert to ENV_VAR if the --uppercase flag is set."""
+    if args.uppercase:
+        return reverse_dns_subdomain(string)
+    elif args.lowercase:
+        return dns_subdomain(string)
+    return string
 
 
 def reverse_dns_subdomain(string):
@@ -204,7 +267,7 @@ def maybe_reverse_dns_subdomain(args, string):
     return reverse_dns_subdomain(string) if args.lowercase else string
 
 
-def kube_init_secret(args, data):
+def kube_init_secret(args, name, data):
     """
     Initialize a Kubernetes secret object (only in memory).
     Data contains the secret data. Each key must consist of alphanumeric
@@ -216,36 +279,51 @@ def kube_init_secret(args, data):
     # https://github.com/kubernetes-incubator/client-python/blob/master/kubernetes/docs/V1Secret.md
     # api_version, data, kind, metadata, string_data, type
     converted_data = {
-        maybe_dns_subdomain(args, key): base64.b64encode(data[key].encode('utf-8')).decode('utf-8')
+        generate_key(args, key): base64.b64encode(data[key].encode('utf-8')).decode('utf-8')
         for key in data
     }
-    metadata = kubernetes.client.V1ObjectMeta(name=args.secret)
+    metadata = kubernetes.client.V1ObjectMeta(name=name)
     return kubernetes.client.V1Secret(data=converted_data, type='Opaque', metadata=metadata)
 
 
-def kube_create_secret(args, data):
+def kube_create_secret(args, namespace, secret, data):
     """ Creates a Kubernetes secret. Returns the api response from Kubernetes."""
     # https://github.com/kubernetes-incubator/client-python/blob/master/kubernetes/docs/CoreV1Api.md#create_namespaced_secret
     kube = kubernetes.client.CoreV1Api()
-    body = kube_init_secret(args, data)
-    return kube.create_namespaced_secret(args.namespace, body)
+    body = kube_init_secret(args, secret, data)
+    return kube.create_namespaced_secret(namespace, body)
 
 
-def kube_replace_secret(args, data):
+def kube_replace_secret(args, namespace, secret, data):
     """ Replaces a kubernetes secret. Returns the api response from Kubernetes. """
     # https://github.com/kubernetes-incubator/client-python/blob/master/kubernetes/docs/CoreV1Api.md#replace_namespaced_secret
     kube = kubernetes.client.CoreV1Api()
-    body = kube_init_secret(args, data)
-    return kube.replace_namespaced_secret(args.secret, args.namespace, body)
+    body = kube_init_secret(args, secret, data)
+    return kube.replace_namespaced_secret(secret, namespace, body)
 
 
-def kube_secret_exists(args):
+def kube_secret_exists(namespace, secret):
     """ Returns True or False if a Kubernetes secret exists or not respectively. """
     # https://github.com/kubernetes-incubator/client-python/blob/master/kubernetes/docs/CoreV1Api.md#read_namespaced_secret
     kube = kubernetes.client.CoreV1Api()
     try:
         # TODO: might be better to call list_namespaced_secrets here.
-        kube.read_namespaced_secret(args.secret, args.namespace)
+        kube.read_namespaced_secret(secret, namespace)
+    except kubernetes.client.rest.ApiException as e:
+        if e.status == 404:
+            return False  # 404 means the secret did not exist, so we can return False
+        else:
+            raise  # don't catch errors you can't resolve.
+    return True
+
+
+def kube_namespace_exists(args):
+    """ Returns True or False if a Kubernetes namespace exists or not respectively. """
+    # https://github.com/kubernetes-incubator/client-python/blob/master/kubernetes/docs/CoreV1Api.md#read_namespaced_secret
+    kube = kubernetes.client.CoreV1Api()
+    try:
+        # TODO: might be better to call list_namespaced_secrets here.
+        kube.read_namespace(args.namespace)
     except kubernetes.client.rest.ApiException as e:
         if e.status == 404:
             return False  # 404 means the secret did not exist, so we can return False
@@ -262,14 +340,14 @@ def kube_read_secret(args):
 
 def kube_read_deployment(args):
     """ Returns the full contents of Kubernetes deployment. """
-    kube = kubernetes.client.AppsV1beta1Api()
+    kube = kubernetes.client.CoreV1Api()
     response = kube.read_namespaced_deployment(args.deployment, args.namespace)
     return response
 
 
 def kube_patch_deployment(args, deployment):
     """ Patches a Kubernetes deployment with data `deployment`. Returns the full contents of the patched deployment. """
-    kube = kubernetes.client.AppsV1beta1Api()
+    kube = kubernetes.client.CoreV1Api()
     return kube.patch_namespaced_deployment(args.deployment, args.namespace, deployment)
 
 
@@ -320,71 +398,116 @@ def init_envs_for_container(args, secrets, container):
     return envs
 
 
-def cmd_inject(args):
-    """
-    Pulls values from a Kubernetes secret and injects them into a deployment as environment variables.
-    There may be multiple containers in a single deployment, so we inject into all of them.
-    """
-    # read the original deployment file
-    deployment = kube_read_deployment(args)
-    # read the secrets
-    secrets = kube_read_secret(args).data
-    # TODO: implement a diff here so we can inform the use if they actually changed anything.
-    # TODO: this is a bit too complex, find a way to simplify
-    data = {
-        'spec': {
-            'template': {
-                'spec': {
-                    'containers': [
-                        {
-                            'name': container.name,
-                            'env': init_envs_for_container(args, secrets, container)
-                        }
-                        for container in deployment.spec.template.spec.containers
-                        # if no --container is passed in, we inject in every container
-                        if (container.name in args.container) or (len(args.container) == 0)
-                    ]
-                }
-            }
-        }
-    }
-    kube_patch_deployment(args, data)
-    print('inject is DEPRECATED; see README.md, use envFrom instead)\n\n'
-          'Injected environment variables into deployment: "{deployment}" '
-          'from secret: "{secret}"'.format(deployment=args.deployment, secret=args.secret))
-
-
 def cmd_push(args):
     """ Pulls values from a Credstash table and stores them in a Kubernetes secret. """
 
     if args.verbose:
         print('checking that "{secret}" exists...'.format(secret=args.secret))
 
-    if kube_secret_exists(args):
+    if kube_secret_exists(args.namespace, args.secret):
         if not args.force:
             print('kubernetes Secret: "{secret}" already exists, run with -f to replace it.'.format(secret=args.secret))
             sys.exit(1)
         else:
             data = credstash_getall(args)
-            kube_replace_secret(args, data)
+            kube_replace_secret(args, args.namespace, args.secret, data)
             print('replaced Kubernetes Secret: "{secret}" with Credstash table: "{table}"'.format(secret=args.secret,
                                                                                                   table=args.table))
     else:
         data = credstash_getall(args)
-        kube_create_secret(args, data)
+        kube_create_secret(args, args.namespace, args.secret, data)
         print('created Kubernetes Secret: "{secret}" with Credstash table: "{table}"'.format(table=args.table,
                                                                                              secret=args.secret))
 
 
-def cmd_daemon(args):
-    # https://boto3.readthedocs.io/en/latest/reference/services/dynamodb.html
-    # https://boto3.readthedocs.io/en/latest/reference/services/dynamodbstreams.html
-    # TODO: what if there are more than 100 streams?
+def cmd_pushall(args):
+    """Syncs a Credstash table with an entire cluster"""
 
-    # -f --force is implied by daemon
-    args.force = True
+    # Map of all namespaces to secrets
+    secretMap = {}
 
-    client = boto3.client('dynamodbstreams')
+    # Depending on whether or not we need to
+    if args.secretname:
+        # If the string contains slashes
+        # we can infer the namespace and secretname
+        _secretname = args.secretname
+        secretkey = args.secretname
+        if "/" in args.secretname:
+            args.namespace, args.secret, _secretname = args.secretname.split('/')
+        else:
+            secretkey = args.namespace + "/" + args.secret + "/" + args.secretname
+
+        if args.verbose:
+            print("Syncing a single secret key: key={key}, ns={namespace}, secret={secret}".format(key=secretkey, namespace=args.namespace, secret=args.secret))
+
+        secrets = {args.secretname: credstash_getone(secretkey, args)}
+
+        secretMap = {
+            args.namespace: set([args.secret])
+        }
+    else:
+        prefix = ''
+
+        # This is incorrect, since it doesn't work for the `default` namespace
+        # TODO: Figure out a way!
+        if args.namespace != 'default':
+            if not kube_namespace_exists(args):
+                print('kubernetes namespace: "{namespace}" doesn\'t exist. Please create it before running kubestash'.format(namespace=args.namespace))
+                sys.exit(1)
+            else:
+                prefix += args.namespace + "/"
+            if args.secret:
+                prefix += args.secret + "/"
+        secrets = credstash_getall(args)
+        secrets = {k: secrets[k] for k in secrets if k.startswith(prefix)}
+
+        # Go through each key and create a dict of namespace->secrets to be created
+        # if they don't exist
+        for secret_key in secrets:
+            ns, secret, env_key = secret_key.split('/')
+            try:
+                secretMap[ns].add(secret)
+            except Exception as e:
+                secretMap[ns] = set()
+                secretMap[ns].add(secret)
+
+    for ns in secretMap:
+        # Maybe fix the method to take just string instead?
+        if not kube_namespace_exists(namedtuple('Arg', ['namespace'])(namespace=ns)):
+            print('kubernetes namespace: "{ns}" doesn\'t exist. Ignoring'.format(ns=ns))
+            continue
+
+        # Iterate through the secrets and make sure they exist
+        for secret in secretMap[ns]:
+            try:
+                prefix = ns + "/" + secret + "/"
+                data = filter_secrets(secrets, ns, secret)
+                if kube_secret_exists(ns, secret):
+                    if args.verbose:
+                        print("Force pushing secret to kubernetes: ns={ns}, secret={secret}".format(ns=ns, secret=secret))
+                    kube_replace_secret(args, ns, secret, data)
+                else:
+                    if args.verbose:
+                        print("Creating and pushing secret to kubernetes: ns={ns}, secret={secret}".format(ns=ns, secret=secret))
+                    kube_create_secret(args, ns, secret, data)
+            except:
+                if args.verbose:
+                    traceback.print_exc()
+                else:
+                    pass
+    if args.verbose:
+        print("All secrets synced")
+
+
+def filter_secrets(secrets, ns, secret):
+    """ Filters the secrets passed and strips away the prefix"""
+    prefix = ns + "/" + secret
+
+    return {k.split('/')[2]: secrets[k] for k in secrets if k.startswith(prefix)}
+
+
+def get_stream_client(args):
+    client = boto3.client('dynamodbstreams', region_name=args.region)
 
     response = client.list_streams(TableName=args.table, Limit=100)
 
@@ -416,6 +539,19 @@ def cmd_daemon(args):
 
     shard_iterator = response['ShardIterator']
 
+    return (client, shard_iterator)
+
+
+def cmd_daemon(args):
+    # https://boto3.readthedocs.io/en/latest/reference/services/dynamodb.html
+    # https://boto3.readthedocs.io/en/latest/reference/services/dynamodbstreams.html
+    # TODO: what if there are more than 100 streams?
+
+    # -f --force is implied by daemon
+    args.force = True
+
+    client, shard_iterator = get_stream_client(args)
+
     cmd_push(args)
 
     if args.verbose:
@@ -442,36 +578,67 @@ def cmd_daemon(args):
         time.sleep(args.interval)
 
 
+def cmd_daemonall(args):
+    args.force = True
+
+    client, shard_iterator = get_stream_client(args)
+
+    if args.verbose:
+        print("checking DynamoDB Stream for changes...")
+
+    response = client.get_records(ShardIterator=shard_iterator, Limit=100)
+
+    while True:
+        shard_iterator = response['NextShardIterator']
+
+        if args.verbose:
+            print("checking DynamoDB Stream for changes...")
+
+        response = client.get_records(ShardIterator=shard_iterator, Limit=100)
+
+        if len(response['Records']) > 0:
+            for record in response['Records']:
+                key = record['dynamodb']['Keys']['name']['S']
+                if args.verbose:
+                    print("detected DynamoDB changes, running push command...")
+                argscopy = copy.copy(args)
+                argscopy.secretname = key
+                cmd_pushall(argscopy)
+
+        time.sleep(args.interval)
+
+
 def main():
     args = parse_args()
 
-    config_file = os.path.expanduser(os.environ.get('KUBECONFIG', '~/.kube/config'))
+    try:
+        kubernetes.config.load_kube_config(context=args.context)
+        contexts, _ = kubernetes.config.list_kube_config_contexts()
+        context_names = [c['name'] for c in contexts]
+        # print a useful error message if the user supplies an invalid context
+        if args.context and args.context not in context_names:
+            print("Kubernetes context '{context}' not found, must be one of: {context_list}"
+                  .format(context=args.context,
+                          context_list=', '.join(context_names)))
+            sys.exit(1)
+    except (ValueError, IOError):
+        # if the KUBECONFIG is missing or invalid, fall back to incluster config
+        kubernetes.config.load_incluster_config()
 
-    if args.verbose:
-        print('loading kubernetes config at: "{config_file}"'.format(config_file=config_file))
-
-    contexts, _ = kubernetes.config.list_kube_config_contexts()
-    context_names = [c['name'] for c in contexts]
-    if args.context and args.context not in context_names:
-        print("Kubernetes context '{context}' not found, must be one of: {context_list}"
-              .format(context=args.context,
-                      context_list=', '.join(context_names)))
-        sys.exit(1)
-
-    kubernetes.config.load_kube_config(config_file=config_file, context=args.context)
-
-
-    # override the host if the user passes in a --proxy
+    # if we're in proxy mode, disable ssl verification
     if args.proxy and (len(args.proxy) == 1):
         kubernetes.client.configuration.host = args.proxy[0]
+        kubernetes.client.configuration.verify_ssl = False
 
     try:
         if args.cmd == 'push':
             cmd_push(args)
-        elif args.cmd == 'inject':
-            cmd_inject(args)
+        elif args.cmd == 'pushall':
+            cmd_pushall(args)
         elif args.cmd == 'daemon':
             cmd_daemon(args)
+        elif args.cmd == 'daemonall':
+            cmd_daemonall(args)
         else:
             pass
     except urllib3.exceptions.MaxRetryError as e:
@@ -505,7 +672,7 @@ def main():
                    '- is your apiserver reachable ?\n\n'
                    'use --proxy HOST to override the host if neccesary\n'
                    ).format(host=kubernetes.client.configuration.host,
-                            config_file=config_file))
+                            config_file=os.path.expanduser(os.environ.get('KUBECONFIG', '~/.kube/config'))))
             sys.exit(1)
         else:
             raise