+++ test/test-teardown.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+set -ex
+cd "$(git rev-parse --show-toplevel)"
+minikube stop
+aws dynamodb delete-table --table-name kubestash