+++ test/test.sh
@@ -1,0 +1,10 @@
+#!/usr/bin/env bash
+set -ex
+cd "$(git rev-parse --show-toplevel)"
+
+./test/test-setup.sh
+./test/test-local.sh
+./test/test-docker.sh
+./test/test-kubernetes.sh
+./test/test-teardown.sh
+